Removed pred_and_pdf_func from models, made it non abstract method in parent class.
